@@ -3,35 +3,35 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@ant-design/compatible": "^1.0.8",
-        "antd": "^4.16.3",
-        "classnames": "^2.3.1",
-        "codemirror": "^5.62.0",
+        "@ant-design/compatible": "^5.1.1",
+        "@ant-design/icons": "^5.0.1",
+        "antd": "^5.3.2",
+        "classnames": "^2.3.2",
+        "codemirror": "^5.65.12",
         "highcharts": "^9.1.2",
         "highcharts-react-official": "^2.2.2",
         "highstock-release": "^6.0.4",
-        "history": "^4.10.1",
-        "jquery": "^3.6.0",
+        "history": "^5.3.0",
+        "jquery": "^3.6.4",
         "mathjs": "^7.6.0",
-        "mergely": "^4.3.1",
+        "mergely": "^4.3.9",
         "moment": "^2.29.4",
-        "npm-check": "^5.9.2",
-        "qs": "^6.10.1",
-        "react": "^16.14.0",
+        "qs": "^6.11.1",
+        "react": "^18.2.0",
+        "react-app-rewired": "^2.2.1",
         "react-day-picker": "^7.4.10",
-        "react-dom": "^16.14.0",
+        "react-dom": "^18.2.0",
         "react-grid-layout": "^0.16.6",
         "react-highcharts": "^16.1.0",
         "react-highlight-words": "^0.16.0",
         "react-jsx-highcharts": "^3.6.1",
         "react-jsx-highstock": "^3.6.1",
-        "react-nl2br": "^0.6.1",
-        "react-router": "^5.2.0",
-        "react-router-dom": "^5.2.0",
+        "react-nl2br": "^1.0.4",
+        "react-router": "^6.10.0",
+        "react-router-dom": "^6.10.0",
+        "react-scripts": "^5.0.1",
         "react-table": "^6.11.5",
-        "sweetalert2": "^10.16.9",
-        "react-app-rewired": "^2.1.8",
-        "react-scripts": "^5.0.1"
+        "sweetalert2": "^11.7.3"
     },
     "scripts": {
         "start": "react-app-rewired start",
@@ -43,6 +43,7 @@
     "proxy": "http://localhost:3001",
     "devDependencies": {
         "cypress": "^8.6.0",
+        "npm-check": "^6.0.1",
         "prettier": "2.5.1"
     },
     "browserslist": [
