@@ -1,50 +1,40 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import { params, getParams } from '../utils/query';
 import { fetchData } from '../utils/Utils';
 import TestBreadcrumb from '../Build/TestBreadcrumb';
 import SearchTestResult from './SearchTestResult';
 import SearchBuildResult from './SearchBuildResult';
 import SearchOutput from './SearchOutput';
-export default class SearchResult extends Component {
-    state = {
-        builds: [],
-        tests: [],
-        parent: null,
-    };
+import { useLocation } from 'react-router-dom';
 
-    async componentDidMount() {
-        await this.updateData();
-    }
+const SearchResult = () => {
+    const [builds, setBuilds] = useState([]);
+    const [tests, setTests] = useState([]);
+    const location = useLocation();
 
-    async componentDidUpdate(prevProps) {
-        if (prevProps.location.search !== this.props.location.search) {
-            await this.updateData();
-        }
-    }
+    useEffect(() => {
+        const updateData = async () => {
+            const { buildId, searchText } = getParams(location.search);
+            const result = await fetchData(
+                `/api/getTestBySearch${params({ buildId, searchText })}`
+            );
 
-    async updateData() {
-        const { buildId, searchText } = getParams(this.props.location.search);
-        const result = await fetchData(
-            `/api/getTestBySearch${params({ buildId, searchText })}`
-        );
+            setBuilds(result.builds);
+            setTests(result.tests);
+        };
+        updateData();
+    }, [location.search]);
 
-        this.setState({
-            builds: result.builds,
-            tests: result.tests,
-        });
-    }
+    const { buildId, searchText } = getParams(location.search);
+    return (
+        <div>
+            <SearchOutput buildId={buildId} />
+            <TestBreadcrumb buildId={buildId} />
+            <SearchTestResult tests={tests} searchText={searchText} />
+            <br />
+            <SearchBuildResult builds={builds} searchText={searchText} />
+        </div>
+    );
+};
 
-    render() {
-        const { builds, tests } = this.state;
-        const { buildId, searchText } = getParams(this.props.location.search);
-        return (
-            <div>
-                <SearchOutput buildId={buildId} />
-                {<TestBreadcrumb buildId={buildId} />}
-                {<SearchTestResult tests={tests} searchText={searchText} />}
-                <br />
-                {<SearchBuildResult builds={builds} searchText={searchText} />}
-            </div>
-        );
-    }
-}
+export default SearchResult;
